+++ src/BroString.h
@@ -35,13 +35,13 @@ public:
 	typedef IdxVec::const_iterator IdxVecCIt;
 
 	// Constructors creating internal copies of the data passed in.
-	BroString(const u_char* str, unsigned int arg_n, int add_NUL);
+	BroString(const u_char* str, uint64_t arg_n, int add_NUL);
 	explicit BroString(const char* str);
 	explicit BroString(const string& str);
 	BroString(const BroString& bs);
 
 	// Constructor that takes owernship of the vector passed in.
-	BroString(int arg_final_NUL, byte_vec str, unsigned int arg_n);
+	BroString(int arg_final_NUL, byte_vec str, uint64_t arg_n);
 
 	BroString();
 	~BroString()	{ Reset(); }
@@ -63,7 +63,7 @@ public:
 	// current contents, if any, and then set the string's
 	// contents to a copy of the string given by the arguments.
 	//
-	void Set(const u_char* str, unsigned int len, int add_NUL=1);
+	void Set(const u_char* str, uint64_t len, int add_NUL=1);
 	void Set(const char* str);
 	void Set(const string& str);
 	void Set(const BroString &str);
@@ -147,7 +147,7 @@ protected:
 	void Reset();
 
 	byte_vec b;
-	unsigned int n;
+	uint64_t n;
 	int final_NUL:1;	// whether we have added a final NUL
 	unsigned int use_free_to_delete:1;	// free() vs. operator delete
 };
